LayoutTests/imported/w3c/web-platform-tests/css/css-ui/parsing/webkit-appearance-alias-expected.txt

CONSOLE MESSAGE: ReferenceError: Can't find variable: test_is_legacy_name_alias

Harness Error (FAIL), message = ReferenceError: Can't find variable: test_is_legacy_name_alias
